Avoid excessive setup overhead; still unable to use record IDs properly

+++ tests/provider.test.setup.js
@@ -12,8 +12,12 @@ const expectedNotes = [
     }
 ];
 
+let currentState;
+
 function addFirstNote() {
-    return noteService.create(expectedNotes[0]);
+    return noteService.deleteAll().then(() => {
+        return noteService.create(expectedNotes[0]);
+    });
 }
 
 function addSecondNote() {
@@ -26,10 +30,15 @@ function doNothing() {
 
 function goToState(desiredState) {
     console.log('Setup state:', desiredState);
+    if(currentState === desiredState) {
+        console.log('Already in state', currentState)
+        return doNothing();
+    }
+    currentState = desiredState; // for next time - assume the state change will succeed
     switch(desiredState) {
         case 'empty': return noteService.deleteAll();
         case 'first note': return addFirstNote();
-        case 'second note': return addSecondNote();
+        case 'second note': return noteService.deleteAll().then(addSecondNote);
         case 'two notes': return addFirstNote().then(addSecondNote);
         default: return doNothing();
     }
@@ -40,7 +49,10 @@ exports.stateService = () => {
         moveToRequestedState: (req, res) => {
             return goToState(req.body.state).then(() => {
                 return res.status(200).send('');
-            })
+            }).catch(err => {
+                currentState = null;
+                return res.status(500).send(err);
+            });
         },
         getExpectedNote: (subscript) => {
             return expectedNotes[subscript];
